// packages/core/src/generated-protocol.ts
export type i32 = number
export type i64 = bigint
export type u16 = number
export type u32 = number
export type u64 = bigint
// ...
export type GuestFilesystemStat = {
    readonly mode: u32
    readonly size: u64
    readonly blocks: u64
    readonly dev: u64
    readonly rdev: u64
    readonly isDirectory: boolean
    readonly isSymbolicLink: boolean
    readonly atimeMs: u64
    readonly mtimeMs: u64
    readonly ctimeMs: u64
    readonly birthtimeMs: u64
    readonly ino: u64
    readonly nlink: u64
    readonly uid: u32
    readonly gid: u32
}
// ...
export enum ProcessSnapshotStatus {
    Running = "Running",
    Exited = "Exited",
    Stopped = "Stopped",
}
// ...
export type ProcessSnapshotEntry = {
    readonly processId: string
    readonly pid: u32
    readonly ppid: u32
    readonly pgid: u32
    readonly sid: u32
    readonly driver: string
    readonly command: string
    readonly args: readonly string[]
    readonly cwd: string
    readonly status: ProcessSnapshotStatus
    readonly exitCode: i32 | null
}
// ...
export type SocketStateEntry = {
    readonly processId: string
    readonly host: string | null
    readonly port: u16 | null
    readonly path: string | null
}